+++ libs/etl/etl/load/daily_web_stats.py
@@ -1,5 +1,6 @@
 from etl.transform.daily_web_stat import DailyWebStat
 from etl.tools.db_manager import DBManager
+from etl.tools.db_structure import TableName
 
 
 def insert_daily_web_stats(web_stats: list[DailyWebStat]):
@@ -7,7 +8,7 @@ def insert_daily_web_stats(web_stats: list[DailyWebStat]):
         return
 
     query = f"""
-        INSERT INTO __SCHEMA_NAME__.daily_web_stats (
+        INSERT INTO __SCHEMA_NAME__.{TableName.DAILY_WEB_STATS} (
             {DailyWebStat.STAT_DATE},
             {DailyWebStat.UNIQUE_VISITORS},
             {DailyWebStat.DETAIL_PAGE_UNIQUE_VISITORS}
