+++ libs/etl/etl/pipelines/companies_pipeline.py
@@ -1,8 +1,6 @@
 from etl.extract.db_queries import get_new_sirets
-from etl.transform.establishment import (
-    keep_valid_sirets,
-    siren_establishment_to_db_establishment,
-)
+from etl.transform.establishment import siren_establishment_to_db_establishment
+from etl.transform.siret_search_error import SiretSearchError
 from etl.extract.siren_extractor import SireneExtractor
 from etl.load.companies import insert_companies, insert_siren_error
 import time
@@ -10,7 +8,7 @@ import time
 class CompaniesPipeline:
     def process_new_sirets(self):
         sirets = get_new_sirets()
-        valid_sirets = keep_valid_sirets(sirets)
+        valid_sirets = SiretSearchError.keep_valid_sirets(sirets)
         if not valid_sirets:
             print("No new SIRETs to process.")
         else:
